move *ArmConnectionStringProvider types into the impl packages
