+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/resourcemanager/AzureServiceBusResourceManagerAutoConfiguration.java
@@ -7,7 +7,7 @@ import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.cloud.autoconfigure.condition.ConditionalOnMissingProperty;
 import com.azure.spring.cloud.autoconfigure.implementation.resourcemanager.ServiceBusResourceMetadata;
 import com.azure.spring.cloud.autoconfigure.implementation.servicebus.properties.AzureServiceBusProperties;
-import com.azure.spring.resourcemanager.connectionstring.ServiceBusArmConnectionStringProvider;
+import com.azure.spring.resourcemanager.implementation.connectionstring.ServiceBusArmConnectionStringProvider;
 import com.azure.spring.resourcemanager.implementation.provisioning.DefaultServiceBusProvisioner;
 import com.azure.spring.resourcemanager.provisioning.ServiceBusProvisioner;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
@@ -52,7 +52,7 @@ public class AzureServiceBusResourceManagerAutoConfiguration extends AzureServic
     @ConditionalOnProperty(prefix = AzureServiceBusProperties.PREFIX, value = "namespace")
     @ConditionalOnMissingProperty(prefix = AzureServiceBusProperties.PREFIX, value = "connection-string")
     @Order
-    public ServiceBusArmConnectionStringProvider serviceBusArmConnectionStringProvider() {
+    ServiceBusArmConnectionStringProvider serviceBusArmConnectionStringProvider() {
         return new ServiceBusArmConnectionStringProvider(this.azureResourceManager,
                                                          this.resourceMetadata,
                                                          this.resourceMetadata.getName());
