+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/eventhubs/AzureEventHubsKafkaAutoConfigurationTests.java
@@ -7,7 +7,7 @@ import com.azure.spring.cloud.autoconfigure.eventhubs.kafka.AzureEventHubsKafkaA
 import com.azure.spring.cloud.autoconfigure.implementation.properties.AzureGlobalProperties;
 import com.azure.spring.core.connectionstring.StaticConnectionStringProvider;
 import com.azure.spring.core.service.AzureServiceType;
-import com.azure.spring.resourcemanager.connectionstring.ArmConnectionStringProvider;
+import com.azure.spring.resourcemanager.implementation.connectionstring.ArmConnectionStringProvider;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
