+++ sdk/spring/spring-cloud-azure-resourcemanager/src/main/java/com/azure/spring/resourcemanager/implementation/connectionstring/ArmConnectionStringProvider.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.resourcemanager.connectionstring;
+package com.azure.spring.resourcemanager.implementation.connectionstring;
 
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.core.connectionstring.ConnectionStringProvider;
