+++ sdk/spring/spring-cloud-azure-resourcemanager/src/test/java/com/azure/spring/resourcemanager/implementation/connectionstring/EventHubsArmConnectionStringProviderTests.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.resourcemanager.connectionstring;
+package com.azure.spring.resourcemanager.implementation.connectionstring;
 
 import com.azure.spring.core.service.AzureServiceType;
