+++ sdk/spring/spring-cloud-azure-resourcemanager/src/test/java/com/azure/spring/resourcemanager/implementation/provisioning/DefaultEventHubsProvisionerTests.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.resourcemanager.provisioning.eventhubs;
+package com.azure.spring.resourcemanager.implementation.provisioning;
 
 import com.azure.core.http.HttpResponse;
 import com.azure.core.management.exception.ManagementException;
@@ -16,7 +16,6 @@ import com.azure.spring.core.properties.resource.AzureResourceMetadata;
 import com.azure.spring.resourcemanager.implementation.crud.EventHubNamespaceCrud;
 import com.azure.spring.resourcemanager.implementation.crud.EventHubsConsumerGroupCrud;
 import com.azure.spring.resourcemanager.implementation.crud.EventHubsCrud;
-import com.azure.spring.resourcemanager.implementation.provisioning.DefaultEventHubsProvisioner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import reactor.util.function.Tuples;
