+++ sdk/spring/spring-cloud-azure-resourcemanager/src/test/java/com/azure/spring/resourcemanager/implementation/provisioning/DefaultServiceBusProvisionerTests.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.resourcemanager.provisioning.servicebus;
+package com.azure.spring.resourcemanager.implementation.provisioning;
 
 import com.azure.core.http.HttpResponse;
 import com.azure.core.management.exception.ManagementException;
@@ -18,7 +18,6 @@ import com.azure.spring.core.properties.resource.AzureResourceMetadata;
 import com.azure.spring.resourcemanager.implementation.crud.ServiceBusQueueCrud;
 import com.azure.spring.resourcemanager.implementation.crud.ServiceBusTopicCrud;
 import com.azure.spring.resourcemanager.implementation.crud.ServiceBusTopicSubscriptionCrud;
-import com.azure.spring.resourcemanager.implementation.provisioning.DefaultServiceBusProvisioner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import reactor.util.function.Tuples;
